PXBF-1643-error-design-updates: move a few of our error handlers out to utilities
